Greenlight... REDLIGHT!
Fixed the hang,  check the change to the SQL Controller to confirm.
Edited out a possible bug in the debug IOWINDOW
Love the look it has, excellent stuff. Just excellent. ^.^
All the printing stuff will be ready to be done the moment i can get to
the competition and debug our debugger.
The new layout looks great ^.^

Starting work on graphing controller, but wanted to commit that fix for
you first :P